Rework handling of transposing instruments (see #1733, #1800)
This patch establishes a consistent default for Score Wizard scores:
All instruments are entered at concert pitch regardless of transposition.

Transposition commands, when needed, are moved from the music block to
the staff definition. This solves several practical problems, and is
also more logically justified since transposition is a notation
convention rather than an inherent property of the instrument or music.