Document that duplicating the list of parts is intentional
I explained this in an earlier review for PR #1810, but it should also
be noted in the code itself. The behavior of the separate \score feature
is based on how I typically set up my own scores, which often end up
with very different part lists for their print and MIDI outputs.