Replace global signals with native Qt versions
As discussed under #1868, I believe Qt's memory management may be
interfering with our custom implementation of global signals and slots.
If this is the case, it could explain the persistent crashes and "object
has been deleted" errors on macOS.

This is intended to test that hypothesis by reimplementing our global
signals using Qt's native signals and slots.