Connect our aboutToQuit signal directly to qApp's
This is an attempt to solve problems particularly associated with macOS
such as crashes on exit (#1868) or "object has been deleted" errors
(#1467 and many others).

My hypothesis is that our non-Qt-based aboutToQuit signal is being
emitted after qApp is garbage collected, so by the time its slots are
called the objects they operate on no longer exist. This patch connects
our signal directly to qApp's so it is instead called before qApp is
destroyed. It also documents to the best of my understanding why we use
a non-Qt-based signal in the first place.

If this works it may avoid the need for a more significant refactor like
I've attempted under #1877.