Make mainwidget an instance attribute so it's not garbage-collected
Adding a widget to a layout changes its parent to the widget on which
the layout is installed. Here the layout is on self.mainwidget, and the
tab bar and view manager become its children. Thus if mainwidget is a
local variable, the tab bar and view manager are garbage-collected when
the constructor returns and mainwidget goes out of scope, even though
they were constructed with the MainWindow instance as their parent.

Hopefully this finally fixes the infamous #1467.