Set an explicit default font on macOS
When I had a chance to briefly test Frescobaldi on a Mac, I noticed the
editor font defaulted to the system sans-serif font. This PR explicitly
sets a monospace default font on macOS to match the behavior on other
platforms.

Do any of our Mac users have a preference for which font? I assume
Monaco is the safest choice since it's been in every macOS version, but
it looks like Apple favors other fonts in recent OS versions.