Add Ctrl+, shortcut to open the Preferences (#1952)
Qt doesn't have a default keyboard shortcut for GNOME and Windows
to open the Preferences¹, so a default shortcut is currently
missing in these platforms.
We can force Ctrl+, which is the default on macOS as well as in
GNOME² and it's also very common on some popular crossplatform
editors.

This is a partial fix for #887

¹ https://doc.qt.io/qt-6/qkeysequence.html#standard-shortcuts
² https://developer.gnome.org/hig/reference/keyboard.html