New icon theme (for macos)
I created an icon theme based on Google Material Icons (with some modifications). I also modified the code so that the app recognizes when switching from light to dark themes (but only on macos, as I don't have the ability to test it on other platforms). I also disabled the system_icons function on Mac, as it was causing conflicts.