Set the maximum zoom for the PDF viewer to 800%
This was suggested as a workaround for qpageview performance issues, but
I believe it is justified anyway since I'm not aware of any practical
need for higher zoom levels in Frescobaldi. I would prefer to make this
change here rather than in qpageview in case other applications using
the latter have more exotic needs.

Fixes #2035.