Icon theme dark/light for MacOS and Windows
I've created an icon theme based on Tabler Icons https://github.com/tabler/tabler-icons (with some modifications). I also modified the code so that the app recognizes when switching from light to dark themes. The theme is activated by checking the 'system_icons' checkbox in the preferences (if unchecked, the app's icon theme will remain the traditional TangoExt). Changing icons in the preferences causes some minor and pre-existing issues (some icons don't update, or, more rarely, the app quits). Once the preference is set, the app is stable. This change should work on Windows (>= 10) and MacOS (but it was tested on MacOS only). It has no effect on Linux platforms: in this case, unlike the other two, the 'system_icons' preference makes sense, and to implement the new theme I would have had to modify the preferences mechanism more invasively (which my limited knowledge of the app prevents me from doing). However, it shouldn't be difficult for a more experienced programmer to add the theme on Linux as well. Fix #1945